+++ CHANGELOG.md
@@ -2,8 +2,10 @@
 
 * Fixed Tank Widget GUI IO reversed
 * Fixed disable recipe logic settings doesn't work
-* Bump up photon version, and fix photon version compatibility
 * Fixed After Recipe Working Event also triggers when structure is destroyed
+* Fixed Multiblock proxy parts don't export pressure
 * Added try-catch to prevent crash when trying to execute kjs scripts
 * Added TransferProxyRecipeEvent to allow customizing transfer proxy recipe behavior
 * Added MBDRecipeType#recipeBuilder to build dynamic recipe as kjs recipe event
+* Added supports that the part can display its controller trait's ui and the controller can display its part's ui
+* Bump up photon version, and fix photon version compatibility